+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/threads/internal/ThreadEntity.kt
@@ -26,7 +26,7 @@ import java.util.Date
  */
 @Suppress("LongParameterList")
 @Entity(tableName = THREAD_ENTITY_TABLE_NAME)
-internal class ThreadEntity(
+internal data class ThreadEntity(
     @PrimaryKey val parentMessageId: String,
     val cid: String,
     val createdByUserId: String,
@@ -40,6 +40,7 @@ internal class ThreadEntity(
     val title: String,
     val read: List<ChannelUserReadEntity>,
     val latestReplyIds: List<String>,
+    val extraData: Map<String, Any>,
 )
 
 internal const val THREAD_ENTITY_TABLE_NAME = "stream_chat_thread"